Refactor rule E0417
