+++ tests/atest/rules/errors/unsupported_setting_in_init_file/__init__.robot
@@ -1,5 +1,5 @@
 *** Settings ***
-Test Template       Template
+Test Template       template
 Default Tags        tag
 Documentation       some documentation
