+++ tests/atest/rules/errors/unsupported_setting_in_init_file/expected_output.txt
@@ -1,0 +1,4 @@
+__init__.robot:2:1:2:14 [E] 0417 Setting 'Test Template' is not supported in initialization files
+__init__.robot:3:1:3:13 [E] 0417 Setting 'Default Tags' is not supported in initialization files
+__init__.resource:2:1:2:14 [E] 0417 Setting 'Test Template' is not supported in initialization files
+__init__.resource:3:1:3:13 [E] 0417 Setting 'Default Tags' is not supported in initialization files
