+++ tests/atest/rules/errors/unsupported_setting_in_init_file/pl_language/expected_output.txt
@@ -1,0 +1,2 @@
+pl_language/__init__.robot:4:1:4:14 [E] 0417 Setting 'Szablon testu' is not supported in initialization files
+pl_language/__init__.resource:4:1:4:14 [E] 0417 Setting 'Szablon testu' is not supported in initialization files
