+++ tests/atest/rules/errors/unsupported_setting_in_init_file/pl_language/test.robot
@@ -1,0 +1,10 @@
+language: pl
+
+*** Ustawienia ***
+Szablon testu       template
+Znaczniki domyślne  tag
+Dokumentacja        some documentation
+
+
+*** Zmienne ***
+${VAR}   a
