+++ tests/atest/rules/errors/unsupported_setting_in_init_file/test.robot
