+++ tests/atest/rules/errors/unsupported_setting_in_init_file/test_rule.py
@@ -1,0 +1,17 @@
+from tests.atest.utils import RuleAcceptance
+
+
+class TestRuleAcceptance(RuleAcceptance):
+    def test_rule(self):
+        self.check_rule(
+            src_files=["__init__.robot", "__init__.resource", "test.robot"],
+            expected_file="expected_output.txt",
+            issue_format="end_col",
+        )
+
+    def test_pl_language(self):
+        self.check_rule(
+            src_files=["pl_language/__init__.robot", "pl_language/__init__.resource", "pl_language/test.robot"],
+            expected_file="pl_language/expected_output.txt",
+            issue_format="end_col",
+        )
