--- tests/atest/rules/naming/unsupported_setting_in_init_file/expected_output.txt
@@ -1,4 +1,0 @@
-__init__.robot:2:1:2:14 [E] 0327 Setting 'Test Template' is not supported in initialization files
-__init__.robot:3:1:3:13 [E] 0327 Setting 'Default Tags' is not supported in initialization files
-__init__.resource:2:1:2:14 [E] 0327 Setting 'Test Template' is not supported in initialization files
-__init__.resource:3:1:3:13 [E] 0327 Setting 'Default Tags' is not supported in initialization files
