--- tests/atest/rules/naming/unsupported_setting_in_init_file/test_rule.py
@@ -1,6 +1,0 @@
-from tests.atest.utils import RuleAcceptance
-
-
-class TestRuleAcceptance(RuleAcceptance):
-    def test_rule(self):
-        self.check_rule(expected_file="expected_output.txt", issue_format="end_col")
